Core hit sound, enemy cap, deleted test scenes (for now)
Implemented a sound for when any enemy hits the core.

The game now has a hard capacity of 31 enemies on screen at once. This feature has been tested reasonably well, HOWEVER: please keep an eye out for bugs involving this. The enemy count is publicly displayed in the EnemyManagement in game object.

Test scenes proved to be too fickle between changes to the core, enemies, and other assets. A different testing features will be implemented soon.